+++ src/test/java/io/github/astrapi69/json/factory/CustomParserFactoryTest.java
@@ -26,11 +26,19 @@ package io.github.astrapi69.json.factory;
 
 import static org.testng.Assert.assertNotNull;
 
+import io.github.astrapi69.json.ClassToJsonSchemaExtensions;
+import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
+/**
+ * The unit test class for the class {@link CustomParserFactory}
+ */
 public class CustomParserFactoryTest
 {
 
+	/**
+	 * Test method for the constructor of {@link CustomParserFactory}
+	 */
 	@Test
 	public void testConstructor()
 	{