+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 ## Change log
 ----------------------
 
+Version 2.2
+-------------
+
+ADDED:
+
+- new factory method that created an object mapper with a jaxb annotation introspector
+- new converter class ClassToJsonSchemaExtensions that can convert a class object to a json schema
+- new dependency jackson-module-jaxb-annotations for processing jaxb annotations
+- new converter class XmlToJsonExtensions that converts xml string to json string
+
+CHANGED:
+
+- update gradle to new version 8.0.1
+- update of com.github.ben-manes.versions.gradle.plugin to new version 0.46.0
+- new gradle-plugin dependency of com.diffplug.spotless:spotless-plugin-gradle in version 6.16.0
+- update of main dependency json to new version 20230227
+- update of main dependency jackson-* to new version 2.14.2
+- update of test dependency silly-collection to new version 21
+
 Version 2.1
 -------------
 
@@ -15,7 +34,7 @@ ADDED:
 CHANGED:
 
 - update gradle to new version 7.6
-- new gradle-plugin dependency of 'com.diffplug.spotless:spotless-plugin-gradle' in version 6.14.0
+- new gradle-plugin dependency of 'com.diffplug.spotless:spotless-plugin-gradle' in version 6.13.0
 - update of com.github.ben-manes.versions.gradle.plugin to new version 0.44.0
 - update of main dependency jackson-* to new version 2.14.1
 - update of main dependency json to new version 20220924